Module websocket::async::futures::prelude

A "prelude" for crates using the futures crate.

This prelude is similar to the standard library's prelude in that you'll almost always want to import its entire contents, but unlike the standard library's prelude you'll have to do so manually. An example of using this is:

use futures::prelude::*;

We may add items to this over time as they become ubiquitous as well, but otherwise this should help cut down on futures-related imports when you're working with the futures crate!

Enums

Async
Return type of future, indicating whether a value is ready or not.

AsyncSink
The result of an asynchronous attempt to send a value to a sink.

Traits

Future
Trait for types which are a placeholder of a value that may become available at some later point in time.

IntoFuture
Class of types which can be converted into a future.

Sink
A Sink is a value into which other values can be sent, asynchronously.

Stream
A stream of values, not all of which may have been produced yet.

Type Definitions

Poll
Return type of the Future::poll method, indicates whether a future's value is ready or not.

StartSend
Return type of the Sink::start_send method, indicating the outcome of a send attempt. See AsyncSink for more details.