Module websocket::client::fragmentUnstable

Helper structs for dealing with WebSocket fragmented messages

The functions string_fragmenter() and binary_fragmenter() provide a tuple of a writer and an iterator. The data is written to the writer, and the iterator is passed to WebSocketClient.frag_send_text() or WebSocketClient.frag_send_bytes(). Since the frag_send functions can only lock the stream for the scope of the function, they must be kept running while the data is being written to the writer. That is, the call to the frag_send function should happen on a different thread to which the data is being written on. See the examples for each function.

Structs

BinaryFragmentIterator
A helper struct for sending fragmented binary data. Use with WebSocketClient.frag_send_bytes().

BinaryFragmentWriter
A helper struct for sending fragmented binary data. Write the data to this object using BinaryFragmentWriter::push().

StringFragmentIterator
A helper struct for sending fragmented strings. Use with WebSocketClient.frag_send_text().

StringFragmentWriter
A helper struct for sending fragmented strings. Write the strings to this object using StringFragmentWriter::push().

Functions

binary_fragmenter
A helper struct for sending fragmented binary data.

Returns a tuple consisting of a writer (to write the data to) and an iterator (to use with WebSocketClient.frag_send_bytes().

use websocket::client::fragment::binary_fragmenter;
use std::thread::Thread;
let (mut writer, iterator) = binary_fragmenter(); //Returns a writer and an iterator
Thread::spawn(move || {
    writer.push(b"This ");
    writer.push(b"is ");
    writer.push(b"a ");
    writer.push(b"binary ");
    writer.push(b"message.");
    writer.finish();            
}).detach();
let _ = client.frag_send_bytes(iterator);

string_fragmenter
A helper struct for sending fragmented strings.

Returns a tuple consisting of a writer (to write the strings to) and an iterator (to use with WebSocketClient.frag_send_text().

use websocket::client::fragment::string_fragmenter;
use std::thread::Thread;
let (mut writer, iterator) = string_fragmenter(); //Returns a writer and an iterator
Thread::spawn(move || {
    writer.push("This ");
    writer.push("is ");
    writer.push("a ");
    writer.push("fragmented ");
    writer.push("message.");
    writer.finish();            
}).detach();
let _ = client.frag_send_text(iterator);