Struct websocket::client::WebSocketClientUnstable

pub struct WebSocketClient<S: DataFrameSender<W>, R: DataFrameReceiver<E>, C: DataFrameConverter<M>, E: Reader, W: Writer, M: WebSocketMessaging> {
    // some fields omitted
}

Represents a WebSocketClient which connects to a WebSocketServer. See the main library documentation for how to obtain a WebSocketClient.

Methods

impl<S: DataFrameSender<W>, R: DataFrameReceiver<E>, C: DataFrameConverter<M>, E: Reader + Send, W: Writer + Send, M: WebSocketMessaging> WebSocketClient<S, R, C, E, W, M>

fn new(sender: S, receiver: R, converter: C) -> WebSocketClient<S, R, C, E, W, M>

Create a WebSocketClient from the specified DataFrameSender and DataFrameReceiver. Not required for normal usage (used internally by WebSocketResponse).

fn send_dataframe(&mut self, dataframe: &WebSocketDataFrame) -> WebSocketResult<()>

Sends a WebSocketDataFrame. Blocks the task until the message has been sent.

fn recv_dataframe(&mut self) -> WebSocketResult<WebSocketDataFrame>

Receives a single WebSocketDataFrame - may corrupt messages received from recv_message(), so do not use both at the same time (ie. either use only recv_dataframe() or use only recv_message())

fn incoming_dataframes(&mut self) -> IncomingDataFrames<S, R, C, E, W, M>

Gets an iterator over incoming data frames

fn incoming_messages(&mut self) -> IncomingMessages<S, R, C, E, W, M>

Gets an iterator over incoming messages.

The iterator always returns Some(), and each iteration will block until a message is received.

for message in client.incoming_messages() {
    match message.unwrap() {
        WebSocketMessage::Text(text) => { println!("Text: {}", text); },
        WebSocketMessage::Binary(data) => { println!("Binary data received"); },
        _ => { }
    }
}

fn frag_send_text<'a, T: ToString>(&'a mut self, text: T) -> WebSocketResult<TextFragmentSender<'a, S, W>>

Sends a fragmented text message by returning a TextFragmentSender

let mut fragment_sender = client.frag_send_text("This").unwrap();
let _ = fragment_sender.send("is ");
let _ = fragment_sender.send("a ");
let _ = fragment_sender.send("fragmented ");
let _ = fragment_sender.finish("message.");

fn frag_send_bytes<'a, T: BytesContainer>(&'a mut self, data: T) -> WebSocketResult<BinaryFragmentSender<'a, S, W>>

Sends a fragmented binary message by returning a BinaryFragmentSender

let mut fragment_sender = client.frag_send_bytes("ascii_bytes").unwrap();
let _ = fragment_sender.send([100u8, ..4].as_slice());
let _ = fragment_sender.finish(vec![4u8, 2, 68, 24]);

fn send_message(&mut self, message: M) -> WebSocketResult<()>

Sends a WebSocketMessage. Blocks the task until the message has been sent.

let message = WebSocketMessage::Text("Hello, server!".to_string());
let _ = client.send_message(message);

fn recv_message(&mut self) -> WebSocketResult<M>

Receives a WebSocketMessage. Blocks the task until a full message is received.

let message = client.recv_message().unwrap();
match message {
    WebSocketMessage::Text(text) => { println!("Text: {}", text); },
    WebSocketMessage::Binary(data) => { println!("Binary data received"); },
    _ => { }
}