Struct websocket::codec::http::HttpClientCodec

pub struct HttpClientCodec;

A codec to be used with tokio codecs that can serialize HTTP requests and deserialize HTTP responses. One can use this on it's own without websockets to make a very bare async HTTP server.

Example

use websocket::async::HttpClientCodec;

let mut core = Core::new().unwrap();
let addr = "crouton.net".parse().unwrap();

let f = TcpStream::connect(&addr, &core.handle())
    .and_then(|s| {
        Ok(s.framed(HttpClientCodec))
    })
    .and_then(|s| {
        s.send(Incoming {
            version: HttpVersion::Http11,
            subject: (Method::Get, RequestUri::AbsolutePath("/".to_string())),
            headers: Headers::new(),
        })
    })
    .map_err(|e| e.into())
    .and_then(|s| s.into_future().map_err(|(e, _)| e))
    .map(|(m, _)| println!("You got a crouton: {:?}", m));

core.run(f).unwrap();