Module websocket::commonUnstable

Contains some common types for use with Rust-WebSocket

Reexports

pub use self::error::WebSocketError;

pub use self::stream::WebSocketStream;

Modules

error
The errors used within Rust-WebSocket

stream
Provides a default implementation of a stream which can be used with Rust-WebSocket.

A single type is provided, capable of carrying either a TcpStream or an SslStream. This is used with convenience methods such as WebSocketRequest::connect(), WebSocketServer::bind()) and WebSocketServer::bind_secure().

Structs

Binary
Represents binary data

Inbound
Represents an inbound object

Local
Represents a local endpoint

Outbound
Represents an outbound object

Remote
Represents a remote endpoint

Text
Represents text data

Type Definitions

WebSocketResult
The type used for WebSocket results