Module websocket::dataframeUnstable

Structs for dealing with WebSocket data frames

Reexports

pub use self::opcode::WebSocketOpcode;

pub use self::sender::{DataFrameSender, WebSocketSender};

pub use self::receiver::{DataFrameReceiver, WebSocketReceiver};

pub use self::converter::{DataFrameConverter, WebSocketConverter};

Modules

converter
Traits and structs related to conversion of data frames into messages

mask
Functions for masking payload data

opcode
Provides an enumeration of WebSocket opcodes

raw
Provides a way to deal with data frames at a lower level

receiver
Traits and structs related to receiving WebSocket data frames

sender
Traits and structs related to sending WebSocket data frames

Structs

WebSocketDataFrame
Represents a WebSocket data frame. The data held in a WebSocketDataFrame is never masked.