Trait websocket::dataframe::converter::WebSocketDataFrameConverterUnstable

pub trait WebSocketDataFrameConverter<M: WebSocketMessaging>: Send {
    fn new() -> Self;
    fn push(&mut self, dataframe: WebSocketDataFrame) -> WebSocketResult<()>;
    fn pop(&mut self) -> Option<M>;

    fn push_all(&mut self, dataframes: &[WebSocketDataFrame]) -> WebSocketResult<()> { ... }
}

A trait which allows the conversion of a series of data frames into a series of messages.

Three functions are required for functionality, and WebSocketConverter is the default implementation.

Required Methods

fn new() -> Self

Creates a new data frame converter.

fn push(&mut self, dataframe: WebSocketDataFrame) -> WebSocketResult<()>

Pushes a data frame on to this converter for processing.

fn pop(&mut self) -> Option<M>

Pops a message from this converter if one is available

Provided Methods

fn push_all(&mut self, dataframes: &[WebSocketDataFrame]) -> WebSocketResult<()>

Pushes a number of data frames on to this converter.