Module websocket::dataframe::receiverUnstable

Traits and structs related to receiving WebSocket data frames

Structs

WebSocketReceiver
The default WebSocket data frame receiver

Traits

DataFrameReceiver
A trait that allows for receiving data frames.

A data frame receiver contains an inner Reader, off of which data frames are read. Data frame receivers read data frames off a Reader, and return WebSocketDataFrames. The function RawDataFrame::read() can be used to read a RawDataFrame from a Reader, enabling the receiver to simply convert that RawDataFrame to a WebSocketDataFrame.