Trait websocket::dataframe::receiver::DataFrameReceiverUnstable

pub trait DataFrameReceiver<R: Reader + Send>: Send {
    fn new(inner: R) -> Self;
    fn recv_dataframe(&mut self) -> WebSocketResult<WebSocketDataFrame>;
}

A trait that allows for receiving data frames.

A data frame receiver contains an inner Reader, off of which data frames are read. Data frame receivers read data frames off a Reader, and return WebSocketDataFrames. The function RawDataFrame::read() can be used to read a RawDataFrame from a Reader, enabling the receiver to simply convert that RawDataFrame to a WebSocketDataFrame.

Required Methods

fn new(inner: R) -> Self

WebSocketDataFrame. Create a new receiver from a Reader

fn recv_dataframe(&mut self) -> WebSocketResult<WebSocketDataFrame>

Reads a data frame from the inner Reader