Trait websocket::dataframe::sender::WebSocketDataFrameSenderUnstable

pub trait WebSocketDataFrameSender<W: Writer + Send>: Send {
    fn new(inner: W) -> Self;
    fn send_dataframe(&mut self, dataframe: &WebSocketDataFrame) -> WebSocketResult<()>;
}

A trait that allows for sending data frames.

A data frame sender contains an inner Writer, on to which data frames are written. Data frame senders convert data frames into raw bytes to be written to a Writer. The function RawDataFrame.write() writes RawDataFrames to a Writer, allowing the data frame sender to simply convert a WebSocketDataFrame into a RawDataFrame and let the RawDataFrame do the writing.

Required Methods

fn new(inner: W) -> Self

Create a new sender from a Writer

fn send_dataframe(&mut self, dataframe: &WebSocketDataFrame) -> WebSocketResult<()>

Write a data frame to the inner Writer