Module websocket::handshake

Structs for WebSocket handshake requests and responses

Structs

HttpVersion
Represents an HTTP version.

WebSocketRequest
Represents a WebSocket handshake request, which is sent from the client to the server. Use the new() function to create a new request, and send it with the WebSocketClient::connect() function. Use the WebSocketClient.receive_handshake_request() method to read a WebSocketRequest from a remote client.

use websocket::handshake::WebSocketRequest;

WebSocketResponse
Represents a WebSocket response which is sent from the server to the client. Use the new() function to create a new response, and send it with the WebSocketClient::send_handshake_response() function.

use websocket::handshake::WebSocketResponse;