Struct websocket::handshake::request::WebSocketRequestUnstable

pub struct WebSocketRequest<R: Reader, W: Writer, B> {
    pub url: RequestUri,
    pub version: HttpVersion,
    pub headers: Headers,
    // some fields omitted
}

Represents a WebSocket request

A WebSocketRequest is used to either make a connection to a WebSocket server using connect() or connect_with(), or received by a server and processed using read() and accept().

A type parameter that is either Inbound or Outbound ensures that only the appropriate methods can be called on either kind of request.

The headers field allows access to the HTTP headers in the request, but short-cut methods are available for accessing common headers.

Fields

url

The target URI for this request.

version

The HTTP version of this request.

headers

The headers of this request.

Methods

impl<R: Reader + Send, W: Writer + Send, B> WebSocketRequest<R, W, B>

fn key(&self) -> Option<&WebSocketKey>

Short-cut to obtain the WebSocketKey value

fn version(&self) -> Option<&WebSocketVersion>

Short-cut to obtain the WebSocketVersion value

fn protocol(&self) -> Option<&WebSocketProtocol>

Short-cut to obtain the WebSocketProtocol value

fn extensions(&self) -> Option<&WebSocketExtensions>

Short-cut to obtain the WebSocketExtensions value

fn origin(&self) -> Option<&Origin>

Short-cut to obtain the Origin value

impl WebSocketRequest<WebSocketStream, WebSocketStream, Outbound>

fn connect(url: Url) -> WebSocketResult<WebSocketRequest<WebSocketStream, WebSocketStream, Outbound>>

Connects to the specified ws:// or wss:// URL.

If a wss:// URL is supplied, a default SslContext is used. connect_ssl_context() can be used to supply an SslContext for use.

fn connect_ssl_context(url: Url, context: &SslContext) -> WebSocketResult<WebSocketRequest<WebSocketStream, WebSocketStream, Outbound>>

Connects to the specified wss:// URL using the given SSL context

impl<R: Reader + Send, W: Writer + Send> WebSocketRequest<R, W, Outbound>

fn new(url: Url, reader: R, writer: W) -> WebSocketResult<WebSocketRequest<R, W, Outbound>>

Create a new outbound request using the specified Reader and Writer.

fn key_mut(&mut self) -> Option<&mut WebSocketKey>

Short-cut to obtain a mutable reference to the WebSocketKey value Note that to add a header that does not already exist, WebSocketRequest.headers.set() must be used.

fn version_mut(&mut self) -> Option<&mut WebSocketVersion>

Short-cut to obtain a mutable reference to the WebSocketVersion value Note that to add a header that does not already exist, WebSocketRequest.headers.set() must be used.

fn protocol_mut(&mut self) -> Option<&mut WebSocketProtocol>

Short-cut to obtaina mutable reference to the WebSocketProtocol value Note that to add a header that does not already exist, WebSocketRequest.headers.set() must be used.

fn extensions_mut(&mut self) -> Option<&mut WebSocketExtensions>

Short-cut to obtain a mutable reference to the WebSocketExtensions value Note that to add a header that does not already exist, WebSocketRequest.headers.set() must be used.

fn origin_mut(&mut self) -> Option<&mut Origin>

Short-cut to obtain a mutable reference to the Origin value Note that to add a header that does not already exist, WebSocketRequest.headers.set() must be used.

fn send(self) -> WebSocketResult<WebSocketResponse<R, W, Inbound>>

Sends the request to the server and returns a response.

impl<R: Reader + Send, W: Writer + Send> WebSocketRequest<R, W, Inbound>

fn read(reader: R, writer: W) -> WebSocketResult<WebSocketRequest<R, W, Inbound>>

Reads an inbound request.

This method is used within servers, and returns an inbound WebSocketRequest. An error will be returned if the request read does not constitute a

fn validate(&self) -> WebSocketResult<()>

Check if this constitutes a valid request.

Note that accept() calls this function internally, however this may be useful for handling bad requests in a custom way.

fn accept(self) -> WebSocketResponse<R, W, Outbound>

Accept this request, ready to send a response.

This function calls validate() on the request, and if the request is found to be invalid, generates a response with a Bad Request status code.

fn fail(self) -> WebSocketResponse<R, W, Outbound>

Fail this request by generating a Bad Request response

fn into_inner(self) -> (R, W)

Return the inner Reader and Writer