Struct websocket::handshake::response::WebSocketResponseUnstable

pub struct WebSocketResponse<R: Reader, W: Writer, B> {
    pub status: StatusCode,
    pub headers: Headers,
    pub version: HttpVersion,
    // some fields omitted
}

Represents a WebSocket response

A WebSocketResponse can sent by a WebSocket server by calling accept() on the WebSocketRequest and then send() on the response, or received by a WebSocket client to be consumed into a WebSocketClient with begin().

A type parameter that is either Inbound or Outbound ensures that only the appropriate methods can be called on either kind of request.

The headers field allows access to the HTTP headers in the response, but short-cut methods are available for accessing common headers.

Fields

status

The status of the response

headers

The headers contained in this response

version

The HTTP version of this response

Methods

impl<R: Reader + Send, W: Writer + Send, B> WebSocketResponse<R, W, B>

fn into_inner(self) -> (R, W)

Return the inner Reader and Writer

fn accept(&self) -> Option<&WebSocketAccept>

Short-cut to obtain the WebSocketAccept value

fn protocol(&self) -> Option<&WebSocketProtocol>

Short-cut to obtain the WebSocketProtocol value

fn extensions(&self) -> Option<&WebSocketExtensions>

Short-cut to obtain the WebSocketExtensions value

impl<R: Reader + Send, W: Writer + Send> WebSocketResponse<R, W, Outbound>

fn new(status: StatusCode, headers: Headers, version: HttpVersion, reader: R, writer: W) -> WebSocketResponse<R, W, Outbound>

Create a new outbound WebSocket response.

fn accept_mut(&mut self) -> Option<&mut WebSocketAccept>

Short-cut to obtain a mutable reference to the WebSocketAccept value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

fn protocol_mut(&mut self) -> Option<&mut WebSocketProtocol>

Short-cut to obtain a mutable reference to the WebSocketProtocol value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

fn extensions_mut(&mut self) -> Option<&mut WebSocketExtensions>

Short-cut to obtain a mutable reference to the WebSocketExtensions value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

fn send_with<A: WebSocketDataFrameSender<W>, B: WebSocketDataFrameReceiver<R>, C: WebSocketDataFrameConverter<M>, M: WebSocketMessaging>(self) -> WebSocketResult<WebSocketClient<A, B, C, R, W, M>>

Send this response with the given WebSocketDataFrameSender, WebSocketDataFrameReceiver and WebSocketDataFrameConverter

fn send_into_inner(self) -> WebSocketResult<(R, W)>

Send this response, retrieving the inner Reader and Writer

fn send(self) -> WebSocketResult<WebSocketClient<WebSocketSender<W, Remote>, WebSocketReceiver<R, Remote>, WebSocketConverter<WebSocketMessage>, R, W, WebSocketMessage>>

Send this response, returning a WebSocketClient ready to transmit/receive data frames

impl<R: Reader + Send, W: Writer + Send> WebSocketResponse<R, W, Inbound>

fn read(reader: R, writer: W) -> WebSocketResult<WebSocketResponse<R, W, Inbound>>

Read an inbound response

fn validate(&self, key: &WebSocketKey) -> WebSocketResult<()>

Check if this response constitutes a successful handshake

fn begin_with<A: WebSocketDataFrameSender<W>, B: WebSocketDataFrameReceiver<R>, C: WebSocketDataFrameConverter<M>, M: WebSocketMessaging>(self) -> WebSocketClient<A, B, C, R, W, M>

Consume this response and return a WebSocketClient ready to transmit/receive data frames using the specified WebSocketDataFrameSender, WebSocketDataFrameReceiver and WebSocketDataFrameConverter.

Does not check if the response was valid. Use validate() to ensure that the response constitutes a successful handshake.

fn begin(self) -> WebSocketClient<WebSocketSender<W, Local>, WebSocketReceiver<R, Local>, WebSocketConverter<WebSocketMessage>, R, W, WebSocketMessage>

Consume this response and return a WebSocketClient ready to transmit/receive data frames