Struct websocket::handshake::WebSocketRequest

pub struct WebSocketRequest {
    pub resource_name: String,
    pub http_version: HttpVersion,
    pub headers: HeaderCollection,
}

Represents a WebSocket handshake request, which is sent from the client to the server. Use the new() function to create a new request, and send it with the WebSocketClient::connect() function. Use the WebSocketClient.receive_handshake_request() method to read a WebSocketRequest from a remote client.

use websocket::handshake::WebSocketRequest;

let protocols = vec!["FirstPreference", "SecondPreference", "ThirdPreference"];
let request = WebSocketRequest::new("ws://www.example.com:1234", protocols.as_slice()).unwrap();
 
assert_eq!(request.host().unwrap().as_slice(), "www.example.com:1234");
assert_eq!(request.version().unwrap().as_slice(), "13");
//...

Fields

resource_name

The resource name of the request. E.g. /path/to/resource for the URI ws://www.example.com/path/to/resource

http_version

The HTTP version of this request

headers

The collection of headers contained in this request

Methods

impl WebSocketRequest

fn new<A: Show>(uri: &str, protocols: &[A]) -> ParseResult<WebSocketRequest>

Creates a new WebSocket handshake request for use with WebSocketClient::connect(). The URI should use the ws:// scheme (wss:// not supported at this time).

fn host(&self) -> Option<String>

Short-cut to get the Host field value of this request

fn connection(&self) -> Option<Vec<String>>

Short-cut to get the Connection field value of this request

fn upgrade(&self) -> Option<Vec<String>>

Short-cut to get the Upgrade field value of this request

fn version(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Version field value of this request

fn key(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Key field value of this request

fn protocol(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Protocol field value of this request

fn extensions(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Extensions field value of this request

fn origin(&self) -> Option<String>

Short-cut to get the Origin field value of this request

impl WebSocketRequest

fn new<A: Show>(uri: &str, protocols: &[A]) -> ParseResult<WebSocketRequest>

Creates a new WebSocket handshake request for use with WebSocketClient::connect(). The URI should use the ws:// scheme (wss:// not supported at this time).

fn host(&self) -> Option<String>

Short-cut to get the Host field value of this request

fn connection(&self) -> Option<Vec<String>>

Short-cut to get the Connection field value of this request

fn upgrade(&self) -> Option<Vec<String>>

Short-cut to get the Upgrade field value of this request

fn version(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Version field value of this request

fn key(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Key field value of this request

fn protocol(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Protocol field value of this request

fn extensions(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Extensions field value of this request

fn origin(&self) -> Option<String>

Short-cut to get the Origin field value of this request