Struct websocket::handshake::WebSocketResponse

pub struct WebSocketResponse {
    pub http_version: HttpVersion,
    pub status_code: uint,
    pub reason_phrase: String,
    pub headers: HeaderCollection,
}

Represents a WebSocket response from a server

Fields

http_version

status_code

The status code of the response (for a successful handshake, this should be 101)

reason_phrase

The human readable reason phrase for the status code (E.g. Switching Protocols)

headers

The collection of headers contained in this WebSocket response

Methods

impl WebSocketResponse

fn new(key: &str, protocol: Option<String>) -> WebSocketResponse

Create a new WebSocket response based on the base-64 encoded key from a client.

fn upgrade(&self) -> Option<String>

fn connection(&self) -> Option<String>

fn accept(&self) -> Option<String>

fn protocol(&self) -> Option<String>

fn version(&self) -> Option<String>

fn extensions(&self) -> Option<Vec<String>>

fn gen_accept<A: ToString>(key: A) -> String

Generates the handshake Sec-WebSocket-Accept value from the given Sec-WebSocket-Key.

fn is_successful<A: ToString>(&self, key: A) -> bool

Returns true if this response indicates a successful handshake

The status code must be 101, the WebSocket-Version must be 13, the Upgrade must be 'websocket', the Connection must be 'Upgrade', the Sec-WebSocket-Accept header must match the expected value from the given key.

impl WebSocketResponse

fn new(key: &str, protocol: Option<String>) -> WebSocketResponse

Create a new WebSocket response based on the base-64 encoded key from a client.

fn upgrade(&self) -> Option<String>

fn connection(&self) -> Option<String>

fn accept(&self) -> Option<String>

fn protocol(&self) -> Option<String>

fn version(&self) -> Option<String>

fn extensions(&self) -> Option<Vec<String>>

fn gen_accept<A: ToString>(key: A) -> String

Generates the handshake Sec-WebSocket-Accept value from the given Sec-WebSocket-Key.

fn is_successful<A: ToString>(&self, key: A) -> bool

Returns true if this response indicates a successful handshake

The status code must be 101, the WebSocket-Version must be 13, the Upgrade must be 'websocket', the Connection must be 'Upgrade', the Sec-WebSocket-Accept header must match the expected value from the given key.