Module websocket::header

Structs representing headers relevant in a WebSocket context.

These headers are commonly used in WebSocket requests and responses. The Header trait from the hyper crate is used.

Re-exports

pub use self::extensions::WebSocketExtensions;

Modules

extensions
Provides the Sec-WebSocket-Extensions header.

parsing
Utility functions for Header implementations.

Structs

Accept
Accept header, defined in RFC7231

AcceptCharset
Accept-Charset header, defined in RFC7231

AcceptEncoding
Accept-Encoding header, defined in RFC7231

AcceptLanguage
Accept-Language header, defined in RFC7231

AcceptRanges
Accept-Ranges header, defined in RFC7233

AccessControlAllowCredentials
Access-Control-Allow-Credentials header, part of CORS

AccessControlAllowHeaders
Access-Control-Allow-Headers header, part of CORS

AccessControlAllowMethods
Access-Control-Allow-Methods header, part of CORS

AccessControlExposeHeaders
Access-Control-Expose-Headers header, part of CORS

AccessControlMaxAge
Access-Control-Max-Age header, part of CORS

AccessControlRequestHeaders
Access-Control-Request-Headers header, part of CORS

AccessControlRequestMethod
Access-Control-Request-Method header, part of CORS

Allow
Allow header, defined in RFC7231

Authorization
Authorization header, defined in RFC7235

Basic
Credential holder for Basic Authentication

Bearer
Token holder for Bearer Authentication, most often seen with oauth

CacheControl
Cache-Control header, defined in RFC7234

Connection
Connection header, defined in RFC7230

ContentDisposition
A Content-Disposition header, (re)defined in RFC6266

ContentEncoding
Content-Encoding header, defined in RFC7231

ContentLanguage
Content-Language header, defined in RFC7231

ContentLength
Content-Length header, defined in RFC7230

ContentRange
Content-Range header, defined in RFC7233

ContentType
Content-Type header, defined in RFC7231

Cookie
Cookie header, defined in RFC6265

Date
Date header, defined in RFC7231

ETag
ETag header, defined in RFC7232

EntityTag
An entity tag, defined in RFC7232

Expires
Expires header, defined in RFC7234

From
From header, defined in RFC7231

HeaderFormatter	Deprecated
A wrapper around any Header with a Display impl that calls fmt_header.

HeaderView
Returned with the HeadersItems iterator.

Headers
A map of header fields on requests and responses.

HeadersItems
An Iterator over the fields in a Headers map.

Host
The Host header.

HttpDate
A time::Time with HTTP formatting and parsing

IfModifiedSince
If-Modified-Since header, defined in RFC7232

IfUnmodifiedSince
If-Unmodified-Since header, defined in RFC7232

LastModified
Last-Modified header, defined in RFC7232

Link
The Link header, defined in RFC5988

LinkValue
A single link-value of a Link header, based on: RFC5988

Location
Location header, defined in RFC7231

Origin
The Origin header.

Origin
Represents an Origin header

Prefer
Prefer header, defined in RFC7240

PreferenceApplied
Preference-Applied header, defined in RFC7240

Protocol
Protocols that appear in the Upgrade header field

Quality
Represents a quality used in quality values.

QualityItem
Represents an item with a quality value as defined in RFC7231.

Referer
Referer header, defined in RFC7231

Server
Server header, defined in RFC7231

SetCookie
Set-Cookie header, defined RFC6265

StrictTransportSecurity
StrictTransportSecurity header, defined in RFC6797

TransferEncoding
Transfer-Encoding header, defined in RFC7230

Upgrade
Upgrade header, defined in RFC7230

UserAgent
User-Agent header, defined in RFC7231

WebSocketAccept
Represents a Sec-WebSocket-Accept header

WebSocketKey
Represents a Sec-WebSocket-Key header.

WebSocketProtocol
Represents a Sec-WebSocket-Protocol header

Enums

AccessControlAllowOrigin
The Access-Control-Allow-Origin response header, part of CORS

ByteRangeSpec
Each Range::Bytes header can contain one or more ByteRangeSpecs. Each ByteRangeSpec defines a range of bytes to fetch

CacheDirective
CacheControl contains a list of these directives.

Charset
A Mime charset.

ConnectionOption
Values that can be in the Connection header.

ContentRangeSpec
Content-Range, described in RFC7233

DispositionParam
A parameter to the disposition type

DispositionType
The implied disposition of the content of the HTTP body

Encoding
A value to represent an encoding used in Transfer-Encoding or Accept-Encoding header.

Expect
The Expect header.

IfMatch
If-Match header, defined in RFC7232

IfNoneMatch
If-None-Match header, defined in RFC7232

IfRange
If-Range header, defined in RFC7233

MediaDesc
A Media Descriptors Enum based on: https://www.w3.org/TR/html401/types.html#h-6.13

Pragma
The Pragma header defined by HTTP/1.0.

Preference
Prefer contains a list of these preferences.

ProtocolName
A protocol name used to identify a spefic protocol. Names are case-sensitive except for the WebSocket value.

Range
Range header, defined in RFC7233

RangeUnit
Range Units, described in RFC7233

ReferrerPolicy
Referrer-Policy header, part of Referrer Policy

RelationType
A Link Relation Type Enum based on: RFC5988

Vary
Vary header, defined in RFC7231

WebSocketVersion
Represents a Sec-WebSocket-Version header

Traits

Header
A trait for any object that will represent a header field and value.

HeaderClone
Internal implementation detail.

HeaderFormat
A trait for any object that will represent a header field and value.

Scheme
An Authorization scheme to be used in the header.

Functions

q
Convenience function to create a Quality from a float.

qitem
Convinience function to wrap a value in a QualityItem Sets q to the default 1.0