Struct websocket::header::ContentType

pub struct ContentType(pub Mime<Vec<(Attr, Value), Global>>);

Content-Type header, defined in RFC7231

The Content-Type header field indicates the media type of the associated representation: either the representation enclosed in the message payload or the selected representation, as determined by the message semantics. The indicated media type defines both the data format and how that data is intended to be processed by a recipient, within the scope of the received message semantics, after any content codings indicated by Content-Encoding are decoded.

ABNF

Content-Type = media-type

Example values

text/html; charset=ISO-8859-4

Examples

use hyper::header::{Headers, ContentType};
use hyper::mime::{Mime, TopLevel, SubLevel};
 
let mut headers = Headers::new();
 
headers.set(
    ContentType(Mime(TopLevel::Text, SubLevel::Html, vec![]))
);

use hyper::header::{Headers, ContentType};
use hyper::mime::{Mime, TopLevel, SubLevel, Attr, Value};
 
let mut headers = Headers::new();
 
headers.set(
    ContentType(Mime(TopLevel::Application, SubLevel::Json,
                     vec![(Attr::Charset, Value::Utf8)]))
);

Tuple Fields
0: Mime<Vec<(Attr, Value), Global>>
Implementations

impl ContentType

pub fn json() -> ContentType

A constructor to easily create a Content-Type: application/json header.

pub fn plaintext() -> ContentType

A constructor to easily create a Content-Type: text/plain; charset=utf-8 header.

pub fn html() -> ContentType

A constructor to easily create a Content-Type: text/html; charset=utf-8 header.

pub fn form_url_encoded() -> ContentType

A constructor to easily create a Content-Type: application/www-form-url-encoded header.

pub fn jpeg() -> ContentType

A constructor to easily create a Content-Type: image/jpeg header.

pub fn png() -> ContentType

A constructor to easily create a Content-Type: image/png header.