Trait websocket::header::Scheme

pub trait Scheme: FromStr + Debug + Clone + Send + Sync {
    fn scheme() -> Option<&'static str>;
    fn fmt_scheme(&self, &mut Formatter<'_>) -> Result<(), Error>;
}

An Authorization scheme to be used in the header.

Required methods

fn scheme() -> Option<&'static str>

An optional Scheme name.

Will be replaced with an associated constant once available.

fn fmt_scheme(&self, &mut Formatter<'_>) -> Result<(), Error>

Format the Scheme data into a header value.