Struct websocket::headers::HeaderCollection

pub struct HeaderCollection {
    // some fields omitted
}

Represents a collection of HTTP headers

Methods

impl HeaderCollection

fn new() -> HeaderCollection

Creates a new HeaderCollection

fn insert<A: ToString, B: ToString>(&mut self, field: A, value: B)

Add the given field-value pair to the collection. If the field is already present, the value is appended to the header using comma-separation. Field names are case-insensitive.

fn contains_field<A: ToString>(&self, field: A) -> bool

Returns true when the specified case-insensitive field name exists in the HeaderCollection.

fn get<A: ToString>(&self, field: A) -> Option<String>

Gets the value of the header with the specified case-insensitive field name.

fn remove<A: ToString>(&mut self, field: A) -> Option<String>

Removes the header with the specified field name from the HeaderCollection.

fn iter(&self) -> Headers

Returns an iterator which iterates over each header in the HeaderCollection.

impl HeaderCollection

fn new() -> HeaderCollection

Creates a new HeaderCollection

fn insert<A: ToString, B: ToString>(&mut self, field: A, value: B)

Add the given field-value pair to the collection. If the field is already present, the value is appended to the header using comma-separation. Field names are case-insensitive.

fn contains_field<A: ToString>(&self, field: A) -> bool

Returns true when the specified case-insensitive field name exists in the HeaderCollection.

fn get<A: ToString>(&self, field: A) -> Option<String>

Gets the value of the header with the specified case-insensitive field name.

fn remove<A: ToString>(&mut self, field: A) -> Option<String>

Removes the header with the specified field name from the HeaderCollection.

fn iter(&self) -> Headers

Returns an iterator which iterates over each header in the HeaderCollection.