Module websocket::message

Structs for WebSocket messages and the transmission of messages

Structs

IncomingMessages
An iterator over incoming messages. Blocks the task and always returns Some.

WebSocketFragmentSerializer
Allows for the serialization of message fragments, to be sent to the remote endpoint. ```no_run use websocket::WebSocketMessage;

// Get a WebSocketFragmentSerializer let mut fragment = sender.fragment();

let message1 = WebSocketMessage::Text("This ".to_string()); let message2 = WebSocketMessage::Text("is ".to_string()); let message3 = WebSocketMessage::Text("a ".to_string()); let message4 = WebSocketMessage::Text("fragmented ".to_string()); let message5 = WebSocketMessage::Text("message.".to_string());

// Send our fragments fragment.send_fragment(&message1); fragment.send_fragment(&message2); fragment.send_fragment(&message3); fragment.send_fragment(&message4);

// Have to tell everyone we're done fragment.finish(&message1);

// Drop this WebSocketFragmentSerializer drop(fragment); ```

WebSocketReceiver
Represents a WebSocket receiver which can receive data from the remote endpoint. All methods are task blocking (but not stream blocking, so you can send and receive concurrently). A WebSocketReceiver can be captured into another task for concurrency

WebSocketSender
Represents a WebSocket sender, capable of transmitting data to the remote endpoint.

use websocket::WebSocketMessage;
 
let mut sender = client.sender(); // Get a sender
let data = "My fancy message".to_string();
let message = WebSocketMessage::Text(data.to_string());
 
let _ = sender.send_message(&message); // Send the message

Enums

WebSocketMessage
Represents a WebSocket message