Struct websocket::message::WebSocketFragmentSerializer

pub struct WebSocketFragmentSerializer<'a> {
    // some fields omitted
}

Allows for the serialization of message fragments, to be sent to the remote endpoint. ```no_run use websocket::WebSocketMessage;

// Get a WebSocketFragmentSerializer let mut fragment = sender.fragment();

let message1 = WebSocketMessage::Text("This ".to_string()); let message2 = WebSocketMessage::Text("is ".to_string()); let message3 = WebSocketMessage::Text("a ".to_string()); let message4 = WebSocketMessage::Text("fragmented ".to_string()); let message5 = WebSocketMessage::Text("message.".to_string());

// Send our fragments fragment.send_fragment(&message1); fragment.send_fragment(&message2); fragment.send_fragment(&message3); fragment.send_fragment(&message4);

// Have to tell everyone we're done fragment.finish(&message1);

// Drop this WebSocketFragmentSerializer drop(fragment); ```

Methods

impl<'a> WebSocketFragmentSerializer<'a>

fn send_fragment(&mut self, message: &WebSocketMessage) -> IoResult<()>

Send a fragment of a message - if the message is finished, use the WebSocketFragmentSerializer::finish() function. Can only be used with a Text or Binary message.

fn finish(&mut self, message: &WebSocketMessage) -> IoResult<()>

Send the final message and tell the remote endpoint the message is complete. Can be used with an empty message if necessary. Can only be used with a Text or Binary message.