Struct websocket::sender::Writer

pub struct Writer<W> {
    pub stream: W,
    pub sender: Sender,
}

A writer that bundles a stream with a serializer to send the messages. This is used in the client's .split() function as the writing component.

It can also be useful to use a websocket connection without a handshake.

Fields
stream: W

The stream that websocket messages will be written to

sender: Sender

The serializer that will be used to serialize the messages

Methods

impl<W> Writer<W> where
W: Write,

pub fn send_dataframe<D>(&mut self, dataframe: &D) -> WebSocketResult<()> where
D: DataFrame,
W: Write,

Sends a single data frame to the remote endpoint.

pub fn send_message<M>(&mut self, message: &M) -> WebSocketResult<()> where
M: Message,

Sends a single message to the remote endpoint.

impl<S> Writer<S> where
S: AsTcpStream + Write,

pub fn shutdown(&self) -> IoResult<()>

Closes the sender side of the connection, will cause all pending and future IO to return immediately with an appropriate value.

pub fn shutdown_all(&self) -> IoResult<()>

Shuts down both Sender and Receiver, will cause all pending and future IO to return immediately with an appropriate value.