Module websocket::server

Provides an implementation of a WebSocket server

Modules

async
The asynchronous implementation of a websocket server.

sync
Provides an implementation of a WebSocket server

upgrade
Allows you to take an existing request or stream of data and convert it into a WebSocket client.

Structs

InvalidConnection
When a sever tries to accept a connection many things can go wrong.

NoTlsAcceptor
Marker struct for a struct not being secure

WsServer
Represents a WebSocket server which can work with either normal (non-secure) connections, or secure WebSocket connections.

Traits

OptionalTlsAcceptor
Trait that is implemented over NoSslAcceptor and SslAcceptor that serves as a generic bound to make a struct with. Used in the Server to specify impls based on whether the server is running over SSL or not.