Type Definition websocket::server::async::Incoming

type Incoming<S> = Box<Stream<Item = (Upgrade<S>, SocketAddr), Error = InvalidConnection<S, BytesMut>>>;

A stream of websocket connections and addresses the server generates.

Each item of the stream is the address of the incoming connection and an Upgrade struct which lets the user decide whether to turn the connection into a websocket connection or reject it.