Struct websocket::server::Connection

pub struct Connection<R: Read, W: Write>(_, _);

Represents a connection to the server that has not been processed yet.

Methods

impl<R: Read, W: Write> Connection<R, W>

fn read_request(self) -> Result<Request<R, W>>

Process this connection and read the request.

impl Connection<WebSocketStream, WebSocketStream>

fn shutdown(&mut self, how: Shutdown) -> Result<()>

Shuts down the currennt connection in the specified way. All future IO calls to this connection will return immediately with an appropriate return value.