Struct websocket::server::InvalidConnection

pub struct InvalidConnection<S, B> where
    S: Stream,  {
    pub stream: Option<S>,
    pub parsed: Option<Request>,
    pub buffer: Option<B>,
    pub error: HyperIntoWsError,
}

When a sever tries to accept a connection many things can go wrong.

This struct is all the information that is recovered from a failed websocket handshake, in case one wants to use the connection for something else (such as HTTP).

Fields
stream: Option<S>

if the stream was successfully setup it will be included here on a failed connection.

parsed: Option<Request>

the parsed request. This is a normal HTTP request meaning you can simply run this server and handle both HTTP and Websocket connections. If you already have a server you want to use, checkout the server::upgrade module to integrate this crate with your server.

buffer: Option<B>

the buffered data that was already taken from the stream

error: HyperIntoWsError

the cause of the failed websocket connection setup