Struct websocket::server::WebSocketServerUnstable

pub struct WebSocketServer<'a> {
    // some fields omitted
}

Represents a WebSocketServer which can work with either normal (non-secure) connections, or secure WebSocket connections.

This is a convenient way to implement WebSocket servers, however it is possible to use any sendable Reader and Writer to obtain a WebSocketClient, so if needed, an alternative server implementation can be used.

Non-secure Servers

extern crate websocket;
use std::thread::Thread;
use std::io::{Listener, Acceptor};
use websocket::{WebSocketServer, WebSocketMessage};

let server = WebSocketServer::bind("127.0.0.1:1234").unwrap();

let mut acceptor = server.listen().unwrap();
for request in acceptor.incoming() {
    // Spawn a new thread for each connection.
    Thread::spawn(move || {
        let request = request.unwrap(); // Get the request
        let response = request.accept(); // Form a response
        let mut client = response.send().unwrap(); // Send the response
        
        let message = WebSocketMessage::Text("Hello, client!".to_string());
        let _ = client.send_message(message);
        
        // ...
    }).detach();
}

Secure Servers

extern crate websocket;
extern crate openssl;
use std::thread::Thread;
use std::io::{Listener, Acceptor};
use websocket::{WebSocketServer, WebSocketMessage};
use openssl::ssl::{SslContext, SslMethod};
use openssl::x509::X509FileType;

let mut context = SslContext::new(SslMethod::Tlsv1).unwrap();
let _ = context.set_certificate_file(&(Path::new("cert.pem")), X509FileType::PEM);
let _ = context.set_private_key_file(&(Path::new("key.pem")), X509FileType::PEM);
let server = WebSocketServer::bind_secure("127.0.0.1:1234", &context).unwrap();

let mut acceptor = server.listen().unwrap();
for request in acceptor.incoming() {
    // Spawn a new thread for each connection.
    Thread::spawn(move || {
        let request = request.unwrap(); // Get the request
        let response = request.accept(); // Form a response
        let mut client = response.send().unwrap(); // Send the response
        
        let message = WebSocketMessage::Text("Hello, client!".to_string());
        let _ = client.send_message(message);
        
        // ...
    }).detach();
}

Custom implementations

Rust-WebSocket allows you to use any (sendable) Reader and Writer with a WebSocketClient. This means you can implement your own server in which ever way you please.

The essential things are to obtain a Reader and a Writer and pass them to WebSocketRequest::read(). This will give you an inbound WebSocketRequest which can be used with WebSocketResponse::new(). Finally, call begin() or begin_with() on the response to obtain a WebSocketClient and begin sending/receiving messages.

Methods

impl<'a> WebSocketServer<'a>

fn bind<T: ToSocketAddr>(addr: T) -> IoResult<WebSocketServer<'a>>

Bind this WebSocketServer to this socket

fn bind_secure<T: ToSocketAddr>(addr: T, context: &'a SslContext) -> IoResult<WebSocketServer<'a>>

Bind this WebSocketServer to this socket, utilising the given SslContext

fn socket_name(&mut self) -> IoResult<SocketAddr>

Get the socket name of this server