Module websocket::server::sync

Provides an implementation of a WebSocket server

Re-exports

pub use server::upgrade::HyperIntoWsError;

pub use server::upgrade::Request;

Type Definitions

AcceptResult

Either the stream was established and it sent a websocket handshake which represents the Ok variant, or there was an error (this is the Err variant).

Server

Represents a WebSocket server which can work with either normal (non-secure) connections, or secure WebSocket connections.