Type Definition websocket::server::sync::AcceptResult

type AcceptResult<S> = Result<Upgrade<S>, InvalidConnection<S, Buffer>>;

Either the stream was established and it sent a websocket handshake which represents the Ok variant, or there was an error (this is the Err variant).