Module websocket::server::upgrade

Allows you to take an existing request or stream of data and convert it into a WebSocket client.

Modules

async

Allows taking an existing stream of data and asynchronously convert it to a websocket client.

sync

Allows you to take an existing request or stream of data and convert it into a WebSocket client.

Structs

WsUpgrade

Intermediate representation of a half created websocket session. Should be used to examine the client’s handshake accept the protocols requested, route the path, etc.

Enums

HyperIntoWsError

Errors that can occur when one tries to upgrade a connection to a websocket connection.

Functions

validate

Check whether an incoming request is a valid WebSocket upgrade attempt.

Type Definitions

Request

A typical request from hyper