Module websocket::server::upgrade::async

Allows taking an existing stream of data and asynchronously convert it to a websocket client.

This module contains the trait that transforms stream into an intermediate struct called Upgrade and the Upgrade struct itself. The Upgrade struct is used to inspect details of the websocket connection (e.g. what protocols it wants to use) and decide whether to accept or reject it.

Traits

IntoWs
Trait to take a stream or similar and attempt to recover the start of a websocket handshake from it (asynchronously). Should be used when a stream might contain a request for a websocket session.

Type Definitions

Upgrade
An asynchronous websocket upgrade.