Enum websocket::server::upgrade::HyperIntoWsError

pub enum HyperIntoWsError {
    MethodNotGet,
    UnsupportedHttpVersion,
    UnsupportedWebsocketVersion,
    NoSecWsKeyHeader,
    NoWsUpgradeHeader,
    NoUpgradeHeader,
    NoWsConnectionHeader,
    NoConnectionHeader,
    Io(Error),
    Parsing(Error),
}

Errors that can occur when one tries to upgrade a connection to a websocket connection.

Variants
MethodNotGet

The HTTP method in a valid websocket upgrade request must be GET

UnsupportedHttpVersion

Currently HTTP 2 is not supported

UnsupportedWebsocketVersion

Currently only WebSocket13 is supported (RFC6455)

NoSecWsKeyHeader

A websocket upgrade request must contain a key

NoWsUpgradeHeader

A websocket upgrade request must ask to upgrade to a websocket

NoUpgradeHeader

A websocket upgrade request must contain an Upgrade header

NoWsConnectionHeader

A websocket upgrade request's Connection header must be Upgrade

NoConnectionHeader

A websocket upgrade request must contain a Connection header

Io(Error)

IO error from reading the underlying socket

Parsing(Error)

Error while parsing an incoming request