Struct websocket::server::upgrade::sync::HyperRequest

pub struct HyperRequest<'a, 'b: 'a>(pub Request<'a, 'b>);

Upgrade a hyper connection to a websocket one.

A hyper request is implicitly defined as a stream from other impls of Stream. Until trait impl specialization comes along, we use this struct to differentiate a hyper request (which already has parsed headers) from a normal stream.

Using this method, one can start a hyper server and check if each request is a websocket upgrade request, if so you can use websockets and hyper on the same port!

use hyper::server::{Server, Request, Response};
use websocket::Message;
use websocket::sync::server::upgrade::IntoWs;
use websocket::sync::server::upgrade::HyperRequest;

Server::http("0.0.0.0:80").unwrap().handle(move |req: Request, res: Response| {
    match HyperRequest(req).into_ws() {
        Ok(upgrade) => {
            // `accept` sends a successful handshake, no need to worry about res
            let mut client = match upgrade.accept() {
                Ok(c) => c,
                Err(_) => panic!(),
            };

            client.send_message(&Message::text("its free real estate"));
        },

        Err((request, err)) => {
            // continue using the request as normal, "echo uri"
            res.send(b"Try connecting over ws instead.").unwrap();
        },
    };
})
.unwrap();