Struct websocket::server::upgrade::sync::RequestStreamPair

pub struct RequestStreamPair<S: Stream>(pub S, pub Request);

If you have your requests separate from your stream you can use this struct to upgrade the connection based on the request given (the request should be a handshake).