Module websocket::stream

Provides the default stream type for WebSocket connections.

Modules

async
A collection of traits and implementations for async streams.

sync
A collection of traits and implementations for synchronous streams.

Structs

ReadWritePair
If you would like to combine an input stream and an output stream into a single stream to talk websockets over then this is the struct for you!

Traits

Stream
Represents a stream that can be read from, and written to. This is an abstraction around readable and writable things to be able to speak websockets over ssl, tcp, unix sockets, etc.