Module websocket::stream::async

A collection of traits and implementations for async streams.

Re-exports

pub use super::ReadWritePair;

Structs

ReadHalf
The readable half of an object returned from AsyncRead::split.

TcpStream
An I/O object representing a TCP stream connected to a remote endpoint.

WriteHalf
The writable half of an object returned from AsyncRead::split.

Traits

AsyncRead
Read bytes asynchronously.

AsyncWrite
Writes bytes asynchronously.

Stream
A stream that can be read from and written to asynchronously. This let's us abstract over many async streams like tcp, ssl, udp, ssh, etc.