Module websocket::stream::sync

A collection of traits and implementations for synchronous streams.

Structs

ReadWritePair

If you would like to combine an input stream and an output stream into a single stream to talk websockets over then this is the struct for you!

TcpStream

A TCP stream between a local and a remote socket.

TlsStream

A stream managing a TLS session.

Enums

Shutdown

Possible values which can be passed to the TcpStream::shutdown method.

Traits

AsTcpStream

The ability access a borrow to an underlying TcpStream, so one can set options on the stream such as nonblocking.

NetworkStream

a Stream that can also be used as a borrow to a TcpStream this is useful when you want to set TcpStream options on a Stream like nonblocking.

Splittable

some streams can be split up into separate reading and writing components TcpStream is an example. This trait marks this ability so one can split up the client into two parts.

Stream

Represents a stream that can be read from, and written to. This is an abstraction around readable and writable things to be able to speak websockets over ssl, tcp, unix sockets, etc.