Struct websocket::WebSocketClient

pub struct WebSocketClient {
    // some fields omitted
}

Represents a WebSocket client.

Methods

impl WebSocketClient

fn connect(request: WebSocketRequest) -> IoResult<WebSocketClient>

Connect to the WebSocket server using the given request. Use WebSocketRequest::new() to create a request for use with this function.

fn from_stream(stream: TcpStream, mask: bool) -> WebSocketClient

Creates a new WebSocketClient from a given TcpStream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false. Typically, WebSocketClient::create() will be used instead of this function.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn send_handshake_response(&mut self, response: WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn receiver(&self) -> WebSocketReceiver

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.

impl WebSocketClient

fn connect(request: WebSocketRequest) -> IoResult<WebSocketClient>

Connect to the WebSocket server using the given request. Use WebSocketRequest::new() to create a request for use with this function.

fn from_stream(stream: TcpStream, mask: bool) -> WebSocketClient

Creates a new WebSocketClient from a given TcpStream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false. Typically, WebSocketClient::create() will be used instead of this function.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn send_handshake_response(&mut self, response: WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn receiver(&self) -> WebSocketReceiver

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.