Module websocket::sync

A collection of handy synchronous-only parts of the crate.

Re-exports

pub use sender;

pub use sender::Writer;

pub use receiver;

pub use receiver::Reader;

pub use server::sync::Server;

pub use client::sync::Client;

Modules

client

A collection of handy synchronous-only parts of the client module.

server

A collection of handy synchronous-only parts of the server module.

stream

A collection of traits and implementations for synchronous streams.

Traits

Stream

Represents a stream that can be read from, and written to. This is an abstraction around readable and writable things to be able to speak websockets over ssl, tcp, unix sockets, etc.