Struct websocket::ws::util::mask::Masker

pub struct Masker<'w, W> where W: Write + 'w {
    // some fields omitted
}

Struct to pipe data into another writer, while masking the data being written

Methods

impl<'w, W> Masker<'w, W> where W: Write + 'w

fn new(key: [u8; 4], endpoint: &'w mut W) -> Self

Create a new Masker with the key and the endpoint to be writter to.