Module websocket::clientUnstable

Structs for dealing with WebSocket clients

Reexports

pub use self::incoming::{IncomingDataFrames, IncomingMessages};

pub use self::fragment::{TextFragmentSender, BinaryFragmentSender};

Modules

fragment
Structs for sending fragmented WebSocket messages

incoming
Provides an iterator over incoming data frames and messages

Structs

WebSocketClient
Represents a WebSocketClient which connects to a WebSocketServer. See the main library documentation for how to obtain a WebSocketClient.