Module websocket::codec::http

Send HTTP requests and responses asynchronously.

This module has both an HttpClientCodec for an async HTTP client and an HttpServerCodec for an async HTTP server.

Structs

HttpClientCodec
A codec to be used with tokio codecs that can serialize HTTP requests and deserialize HTTP responses. One can use this on it's own without websockets to make a very bare async HTTP server.

HttpServerCodec
A codec that can be used with streams implementing AsyncRead + AsyncWrite that can serialize HTTP responses and deserialize HTTP requests. Using this with an async TcpStream will give you a very bare async HTTP server.

Enums

HttpCodecError
Any error that can happen during the writing or parsing of HTTP requests and responses. This consists of HTTP parsing errors (the Http variant) and errors that can occur when writing to IO (the Io variant).