Struct websocket::dataframe::DataFrame

pub struct DataFrame {
    pub finished: bool,
    pub reserved: [bool; 3],
    pub opcode: Opcode,
    pub data: Vec<u8>,
}

Represents a WebSocket data frame.

The data held in a DataFrame is never masked. Masking/unmasking is done when sending and receiving the data frame,

This DataFrame, unlike the standard Message implementation (which also implements the DataFrame trait), owns its entire payload. This means that calls to payload don't allocate extra memory (again unlike the default Message implementation).

Fields
finished: bool

Whether or no this constitutes the end of a message

reserved: [bool; 3]

The reserved portion of the data frame (RFC6455 5.2)

opcode: Opcode

The opcode associated with this data frame

data: Vec<u8>

The payload associated with this data frame

Methods

impl DataFrame

pub fn new(finished: bool, opcode: Opcode, data: Vec<u8>) -> DataFrame

Creates a new DataFrame.

pub fn read_dataframe_body(
header: DataFrameHeader,
body: Vec<u8>,
should_be_masked: bool
) -> WebSocketResult<Self>

Take the body and header of a dataframe and combine it into a single Dataframe struct. A websocket message can be made up of many individual dataframes, use the methods from the Message or OwnedMessage structs to take many of these and create a websocket message.

pub fn read_dataframe<R>(
reader: &mut R,
should_be_masked: bool
) -> WebSocketResult<Self> where
R: Read,

Reads a DataFrame from a Reader.