Struct websocket::handshake::WebSocketResponse

pub struct WebSocketResponse {
    pub http_version: HttpVersion,
    pub status_code: uint,
    pub reason_phrase: String,
    pub headers: HeaderCollection,
}

Represents a WebSocket response which is sent from the server to the client. Use the new() function to create a new response, and send it with the WebSocketClient::send_handshake_response() function.

use websocket::handshake::WebSocketResponse;

let response = WebSocketResponse::new("dGhlIHNhbXBsZSBub25jZQ==", Some("myProtocol"));
 
assert_eq!(response.accept().unwrap().as_slice(), "s3pPLMBiTxaQ9kYGzzhZRbK+xOo=");
assert_eq!(response.upgrade().unwrap().as_slice(), "websocket");
// ...

Note that we can also send WebSocket failure responses.

use websocket::handshake::WebSocketResponse;
 
let response = WebSocketResponse::failure(403, "Forbidden");
assert_eq!(response.status_code, 403);
assert_eq!(response.reason_phrase.as_slice(), "Forbidden");
 
// ...

Fields

http_version

The HTTP version of this request

status_code

The status code of the response (for a successful handshake, this should be 101)

reason_phrase

The human readable reason phrase for the status code (E.g. Switching Protocols)

headers

The collection of headers contained in this WebSocket response

Methods

impl WebSocketResponse

fn new<A: Show>(key: &str, protocol: Option<A>) -> WebSocketResponse

Create a new WebSocket response based on the base-64 encoded key from a client.

fn failure<A: ToString>(status_code: uint, reason_phrase: A) -> WebSocketResponse

Create a WebSocket response with a particular status code and reason phrase - generally used to indicate a handshake failure.

fn upgrade(&self) -> Option<String>

Short-cut to get the Upgrade field value of this request

fn connection(&self) -> Option<String>

Short-cut to get the Connection field value of this request

fn accept(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Accept field value of this request

fn protocol(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Protocol field value of this request

fn version(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Version field value of this request This may be present when the handshake fails (it should not appear on a successful handshake response.

fn extensions(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Extensions field value of this request

fn gen_accept<A: ToString>(key: A) -> String

Generates the handshake Sec-WebSocket-Accept value from the given Sec-WebSocket-Key.

fn is_successful<A: ToString>(&self, key: A) -> bool

Returns true if this response indicates a successful handshake

The status code must be 101, the WebSocket-Version must be 13, the Upgrade must be 'websocket', the Connection must be 'Upgrade', the Sec-WebSocket-Accept header must match the expected value from the given key.

impl WebSocketResponse

fn new<A: Show>(key: &str, protocol: Option<A>) -> WebSocketResponse

Create a new WebSocket response based on the base-64 encoded key from a client.

fn failure<A: ToString>(status_code: uint, reason_phrase: A) -> WebSocketResponse

Create a WebSocket response with a particular status code and reason phrase - generally used to indicate a handshake failure.

fn upgrade(&self) -> Option<String>

Short-cut to get the Upgrade field value of this request

fn connection(&self) -> Option<String>

Short-cut to get the Connection field value of this request

fn accept(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Accept field value of this request

fn protocol(&self) -> Option<String>

Short-cut to get the Sec-WebSocket-Protocol field value of this request

fn version(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Version field value of this request This may be present when the handshake fails (it should not appear on a successful handshake response.

fn extensions(&self) -> Option<Vec<String>>

Short-cut to get the Sec-WebSocket-Extensions field value of this request

fn gen_accept<A: ToString>(key: A) -> String

Generates the handshake Sec-WebSocket-Accept value from the given Sec-WebSocket-Key.

fn is_successful<A: ToString>(&self, key: A) -> bool

Returns true if this response indicates a successful handshake

The status code must be 101, the WebSocket-Version must be 13, the Upgrade must be 'websocket', the Connection must be 'Upgrade', the Sec-WebSocket-Accept header must match the expected value from the given key.