Module websocket::header

Structs representing headers relevant in a WebSocket context.

These headers are commonly used in WebSocket requests and responses. The Header trait from the hyper crate is used.

Reexports

pub use self::extensions::WebSocketExtensions;

Modules

extensions
Provides the Sec-WebSocket-Extensions header.

Structs

Headers
A map of header fields on requests and responses.

Origin
Represents an Origin header

WebSocketAccept
Represents a Sec-WebSocket-Accept header

WebSocketKey
Represents a Sec-WebSocket-Key header.

WebSocketProtocol
Represents a Sec-WebSocket-Protocol header

Enums

WebSocketVersion
Represents a Sec-WebSocket-Version header