Trait websocket::header::HeaderFormat

pub trait HeaderFormat: Debug + HeaderClone + Any + Typeable + Send + Sync {
    fn fmt_header(&self, f: &mut Formatter<'_>) -> Result<(), Error>;
}

A trait for any object that will represent a header field and value.

This trait represents the formatting of a Header for output to a TcpStream.

Required methods

fn fmt_header(&self, f: &mut Formatter<'_>) -> Result<(), Error>

Format a header to be output into a TcpStream.

This method is not allowed to introduce an Err not produced by the passed-in Formatter.