Module websocket::message

Structs for WebSocket messages and the transmission of messages

Structs

IncomingMessages
An iterator over incoming messages. Blocks the task and always returns Some.

WebSocketFragmentSerializer
Allows for the serialization of message fragments, to be sent to the remote endpoint. Any number of messages can be sent (but only text or binary messages, as control messages can never be fragmented) using the send_fragment() method, and when the final fragment is to be sent, use the finish() method. After calling the finish method, the remote endpoint will assemble all of the received message fragments into a single message containing all of the fragment data concatenated together.

use websocket::message::WebSocketMessage;
// Get a WebSocketFragmentSerializer
let mut fragment = sender.fragment();
 
let message1 = WebSocketMessage::Text("This ".to_string());
let message2 = WebSocketMessage::Text("is ".to_string());
let message3 = WebSocketMessage::Text("a ".to_string());
let message4 = WebSocketMessage::Text("fragmented ".to_string());
let message5 = WebSocketMessage::Text("message.".to_string());
 
// Send our fragments
fragment.send_fragment(&message1); // Each one is sent immediately
fragment.send_fragment(&message2);
fragment.send_fragment(&message3);
fragment.send_fragment(&message4);
 
// Now tell them we're done
fragment.finish(&message1); // Now they'll assemble the fragments into a single message
 
// Drop this WebSocketFragmentSerializer, or let it go out of scope
drop(fragment);

WebSocketReceiver
Represents a WebSocket receiver which can receive data from the remote endpoint. All methods are task blocking (but not stream blocking, so you can send and receive concurrently). A WebSocketReceiver can be captured into another task for concurrency

WebSocketSender
Represents a WebSocket sender, capable of transmitting data to the remote endpoint. Use the send_message() method to send a single, whole message. If you need to send a series of message fragments (e.g. if you need to send a message of unknown length), use the fragment() method to obtain a WebSocketFragmentSerializer, which can be used to do so.

use websocket::message::WebSocketMessage;
 
let mut sender = client.sender(); // Get a sender
let data = "My fancy message".to_string();
let message = WebSocketMessage::Text(data.to_string());
 
let _ = sender.send_message(&message); // Send the message

Enums

WebSocketMessage
Represents a WebSocket message. When receiving messages, the resulting messages are always entire messages (never fragments). You are able to, however, send messages as fragments using a WebSocketFragmentSerializer. These messages will be sent to the remote endpoint and reassembled into a single message.