Enum websocket::message::WebSocketMessage

pub enum WebSocketMessage {
    Text(String),
    Binary(Vec<u8>),
    Close(Vec<u8>),
    Ping(Vec<u8>),
    Pong(Vec<u8>),
}

Represents a WebSocket message. When receiving messages, the resulting messages are always entire messages (never fragments). You are able to, however, send messages as fragments using a WebSocketFragmentSerializer. These messages will be sent to the remote endpoint and reassembled into a single message.

Variants

Text

A message containing UTF-8 text data

Binary

A message containing binary data

Close

A message which indicates closure of the WebSocket connection. This message may or may not contain data

Ping

A ping message - should be responded to with a pong message. Usually the pong message will be sent with the same data as the received ping message.

Pong

A pong message