Struct websocket::message::WebSocketFragmentSerializer

pub struct WebSocketFragmentSerializer<'a, S: 'a + Stream + Clone> {
    // some fields omitted
}

Allows for the serialization of message fragments, to be sent to the remote endpoint. Any number of messages can be sent (but only text or binary messages, as control messages can never be fragmented) using the send_fragment() method, and when the final fragment is to be sent, use the finish() method. After calling the finish method, the remote endpoint will assemble all of the received message fragments into a single message containing all of the fragment data concatenated together.

use websocket::message::WebSocketMessage;
// Get a WebSocketFragmentSerializer
let mut fragment = sender.fragment();
 
let message1 = WebSocketMessage::Text("This ".to_string());
let message2 = WebSocketMessage::Text("is ".to_string());
let message3 = WebSocketMessage::Text("a ".to_string());
let message4 = WebSocketMessage::Text("fragmented ".to_string());
let message5 = WebSocketMessage::Text("message.".to_string());
 
// Send our fragments
fragment.send_fragment(&message1); // Each one is sent immediately
fragment.send_fragment(&message2);
fragment.send_fragment(&message3);
fragment.send_fragment(&message4);
 
// Now tell them we're done
fragment.finish(&message1); // Now they'll assemble the fragments into a single message
 
// Drop this WebSocketFragmentSerializer, or let it go out of scope
drop(fragment);

Methods

impl<'a, S: Stream + Clone> WebSocketFragmentSerializer<'a, S>

fn send_fragment(&mut self, message: &WebSocketMessage) -> IoResult<()>

Send a fragment of a message - if the message is finished, use the WebSocketFragmentSerializer::finish() function. Can only be used with a Text or Binary message.

fn finish(&mut self, message: &WebSocketMessage) -> IoResult<()>

Send the final message and tell the remote endpoint the message is complete. Can be used with an empty message if necessary. Can only be used with a Text or Binary message.

Once this has been called, you can send a new fragmented message using the send_fragment() method.