Struct websocket::receiver::Reader

pub struct Reader<R> where
    R: Read,  {
    pub stream: BufReader<R>,
    pub receiver: Receiver,
}

This reader bundles an existing stream with a parsing algorithm. It is used by the client in its .split() function as the reading component.

Fields
stream: BufReader<R>

the stream to be read from

receiver: Receiver

the parser to parse bytes into messages

Methods

impl<R> Reader<R> where
R: Read,

fn recv_dataframe(&mut self) -> WebSocketResult<DataFrame>

Reads a single data frame from the remote endpoint.

fn incoming_dataframes<'a>(
&'a mut self
) -> DataFrameIterator<'a, Receiver, BufReader<R>>

Returns an iterator over incoming data frames.

fn recv_message<'m, M, D, I>(&mut self) -> WebSocketResult<M> where
M: Message<'m, D, DataFrameIterator = I>,
I: Iterator<Item = D>,
D: DataFrameable,

Reads a single message from this receiver.

fn incoming_messages<'a, M, D>(
&'a mut self
) -> MessageIterator<'a, Receiver, D, M, BufReader<R>> where
M: Message<'a, D>,
D: DataFrameable,

An iterator over incoming messsages. This iterator will block until new messages arrive and will never halt.

impl<S> Reader<S> where
S: AsTcpStream + Stream + Read,

fn shutdown(&self) -> IoResult<()>

Closes the receiver side of the connection, will cause all pending and future IO to return immediately with an appropriate value.

fn shutdown_all(&self) -> IoResult<()>

Shuts down both Sender and Receiver, will cause all pending and future IO to return immediately with an appropriate value.