Struct websocket::receiver::Reader

pub struct Reader<R> where
    R: Read,  {
    pub stream: BufReader<R>,
    pub receiver: Receiver,
}

This reader bundles an existing stream with a parsing algorithm. It is used by the client in its .split() function as the reading component.

Fields
stream: BufReader<R>

the stream to be read from

receiver: Receiver

the parser to parse bytes into messages

Methods

impl<R> Reader<R> where
R: Read,

pub fn recv_dataframe(&mut self) -> WebSocketResult<DataFrame>

Reads a single data frame from the remote endpoint.

Important traits for DataFrameIterator<'a, Recv, R>
impl<'a, Recv, R> Iterator for DataFrameIterator<'a, Recv, R> where
Recv: 'a + Receiver,
R: Read, type Item = WebSocketResult<Recv::F>;

pub fn incoming_dataframes(
&mut self
) -> DataFrameIterator<Receiver, BufReader<R>>

Returns an iterator over incoming data frames.

pub fn recv_message<I>(&mut self) -> WebSocketResult<OwnedMessage> where
I: Iterator<Item = DataFrame>,

Reads a single message from this receiver.

Important traits for MessageIterator<'a, Recv, R>
impl<'a, Recv, R> Iterator for MessageIterator<'a, Recv, R> where
Recv: 'a + Receiver,
R: Read, type Item = WebSocketResult<Recv::M>;

pub fn incoming_messages<'a>(
&'a mut self
) -> MessageIterator<'a, Receiver, BufReader<R>>

An iterator over incoming messsages. This iterator will block until new messages arrive and will never halt.

impl<S> Reader<S> where
S: AsTcpStream + Stream + Read,

pub fn shutdown(&self) -> IoResult<()>

Closes the receiver side of the connection, will cause all pending and future IO to return immediately with an appropriate value.

pub fn shutdown_all(&self) -> IoResult<()>

Shuts down both Sender and Receiver, will cause all pending and future IO to return immediately with an appropriate value.