Struct websocket::receiver::Reader

pub struct Reader<R> where
    R: Read,  {
    pub stream: BufReader<R>,
    pub receiver: Receiver,
}

This reader bundles an existing stream with a parsing algorithm. It is used by the client in its .split() function as the reading component.

Fields
stream: BufReader<R>

the stream to be read from

receiver: Receiver

the parser to parse bytes into messages

Implementations

impl<R> Reader<R> where
R: Read,

pub fn recv_dataframe(&mut self) -> WebSocketResult<DataFrame>

Reads a single data frame from the remote endpoint.

pub fn incoming_dataframes(
&mut self
) -> DataFrameIterator<'_, Receiver, BufReader<R>>

Returns an iterator over incoming data frames.

pub fn recv_message(&mut self) -> WebSocketResult<OwnedMessage>

Reads a single message from this receiver.

pub fn incoming_messages<'a>(
&'a mut self
) -> MessageIterator<'a, Receiver, BufReader<R>>

An iterator over incoming messsages. This iterator will block until new messages arrive and will never halt.

impl<S> Reader<S> where
S: AsTcpStream + Stream + Read,

pub fn shutdown(&self) -> IoResult<()>

Closes the receiver side of the connection, will cause all pending and future IO to return immediately with an appropriate value.

pub fn shutdown_all(&self) -> IoResult<()>

Shuts down both Sender and Receiver, will cause all pending and future IO to return immediately with an appropriate value.