Struct websocket::server::response::Response

pub struct Response<R: Read, W: Write> {
    pub status: StatusCode,
    pub headers: Headers,
    pub version: HttpVersion,
    // some fields omitted
}

Represents a server-side (outgoing) response.

Fields
status: StatusCode

The status of the response

headers: Headers

The headers contained in this response

version: HttpVersion

The HTTP version of this response

Methods

impl<R: Read, W: Write> Response<R, W>

pub fn accept(&self) -> Option<&WebSocketAccept>

Short-cut to obtain the WebSocketAccept value

pub fn protocol(&self) -> Option<&WebSocketProtocol>

Short-cut to obtain the WebSocketProtocol value

pub fn extensions(&self) -> Option<&WebSocketExtensions>

Short-cut to obtain the WebSocketExtensions value

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_reader(&self) -> &R

Returns a reference to the inner Reader.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_writer(&self) -> &W

Returns a reference to the inner Writer.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut_reader(&mut self) -> &mut R

Returns a mutable reference to the inner Reader.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut_writer(&mut self) -> &mut W

Returns a mutable reference to the inner Writer.

pub fn get_request(&self) -> &Request<R, W>

Returns a reference to the request associated with this response/

pub fn into_inner(self) -> (R, W)

Return the inner Reader and Writer

pub fn new(request: Request<R, W>) -> Response<R, W>

Create a new outbound WebSocket response.

pub fn bad_request(request: Request<R, W>) -> Response<R, W>

Create a Bad Request response

pub fn accept_mut(&mut self) -> Option<&mut WebSocketAccept>

Short-cut to obtain a mutable reference to the WebSocketAccept value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

pub fn protocol_mut(&mut self) -> Option<&mut WebSocketProtocol>

Short-cut to obtain a mutable reference to the WebSocketProtocol value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

pub fn extensions_mut(&mut self) -> Option<&mut WebSocketExtensions>

Short-cut to obtain a mutable reference to the WebSocketExtensions value Note that to add a header that does not already exist, WebSocketResponse.headers.set() must be used.

pub fn send_with<D, B, C>(
self,
sender: B,
receiver: C
) -> WebSocketResult<Client<D, B, C>> where
B: Sender,
C: Receiver<D>,
D: DataFrameable,

Send this response with the given data frame type D, Sender B and Receiver C.

pub fn send_into_inner(self) -> WebSocketResult<(R, W)>

Send this response, retrieving the inner Reader and Writer

pub fn send(self) -> WebSocketResult<Client<DataFrame, Sender<W>, Receiver<R>>>

Send this response, returning a Client ready to transmit/receive data frames