Struct websocket::server::Server

pub struct Server<'a> {
    // some fields omitted
}

Represents a WebSocket server which can work with either normal (non-secure) connections, or secure WebSocket connections.

This is a convenient way to implement WebSocket servers, however it is possible to use any sendable Reader and Writer to obtain a WebSocketClient, so if needed, an alternative server implementation can be used.

Non-secure Servers

extern crate websocket;
use std::thread::Thread;
use std::old_io::{Listener, Acceptor};
use websocket::{Server, Message};

let server = Server::bind("127.0.0.1:1234").unwrap();

let mut acceptor = server.listen().unwrap();
for request in acceptor.incoming() {
    // Spawn a new thread for each connection.
    Thread::spawn(move || {
        let request = request.unwrap(); // Get the request
        let response = request.accept(); // Form a response
        let mut client = response.send().unwrap(); // Send the response
        
        let message = Message::Text("Hello, client!".to_string());
        let _ = client.send_message(message);
        
        // ...
    });
}

Secure Servers

extern crate websocket;
extern crate openssl;
use std::thread::Thread;
use std::old_io::{Listener, Acceptor};
use websocket::{Server, Message};
use openssl::ssl::{SslContext, SslMethod};
use openssl::x509::X509FileType;

let mut context = SslContext::new(SslMethod::Tlsv1).unwrap();
let _ = context.set_certificate_file(&(Path::new("cert.pem")), X509FileType::PEM);
let _ = context.set_private_key_file(&(Path::new("key.pem")), X509FileType::PEM);
let server = Server::bind_secure("127.0.0.1:1234", &context).unwrap();

let mut acceptor = server.listen().unwrap();
for request in acceptor.incoming() {
    // Spawn a new thread for each connection.
    Thread::spawn(move || {
        let request = request.unwrap(); // Get the request
        let response = request.accept(); // Form a response
        let mut client = response.send().unwrap(); // Send the response
        
        let message = Message::Text("Hello, client!".to_string());
        let _ = client.send_message(message);
        
        // ...
    });
}

Methods

impl<'a> Server<'a>

fn bind<T: ToSocketAddr>(addr: T) -> IoResult<Server<'a>>

Bind this Server to this socket

fn bind_secure<T: ToSocketAddr>(addr: T, context: &'a SslContext) -> IoResult<Server<'a>>

Bind this Server to this socket, utilising the given SslContext

fn socket_name(&mut self) -> IoResult<SocketAddr>

Get the socket name of this server