Trait websocket::server::OptionalTlsAcceptor

pub trait OptionalTlsAcceptor { }

Trait that is implemented over NoSslAcceptor and SslAcceptor that serves as a generic bound to make a struct with. Used in the Server to specify impls based on whether the server is running over SSL or not.