Module websocket::server::upgrade

Allows you to take an existing request or stream of data and convert it into a WebSocket client.

Modules

from_hyper
Upgrade a hyper connection to a websocket one.

Structs

Buffer
This crate uses buffered readers to read in the handshake quickly, in order to interface with other use cases that don't use buffered readers the buffered readers is deconstructed when it is returned to the user and given as the underlying reader and the buffer.

RequestStreamPair
If you have your requests separate from your stream you can use this struct to upgrade the connection based on the request given (the request should be a handshake).

WsUpgrade
Intermediate representation of a half created websocket session. Should be used to examine the client's handshake accept the protocols requested, route the path, etc.

Enums

HyperIntoWsError
Errors that can occur when one tries to upgrade a connection to a websocket connection.

Traits

IntoWs
Trait to take a stream or similar and attempt to recover the start of a websocket handshake from it. Should be used when a stream might contain a request for a websocket session.

Type Definitions

Request
A typical request from hyper