Struct websocket::stream::ReadWritePair

pub struct ReadWritePair<R, W>(pub R, pub W)
 where
    R: Read,
    W: Write;

If you would like to combine an input stream and an output stream into a single stream to talk websockets over then this is the struct for you!

This is useful if you want to use different mediums for different directions.