Trait websocket::stream::sync::AsTcpStream

pub trait AsTcpStream {
    fn as_tcp(&self) -> &TcpStream;
}

The ability access a borrow to an underlying TcpStream, so one can set options on the stream such as nonblocking.

Required methods

fn as_tcp(&self) -> &TcpStream

Get a borrow of the TcpStream

Loading content...