Trait websocket::stream::sync::AsTcpStream

pub trait AsTcpStream {
    fn as_tcp(&self) -> &TcpStream;
}

The ability access a borrow to an underlying TcpStream, so one can set options on the stream such as nonblocking.

Required Methods

Important traits for TcpStream
impl Write for TcpStreamimpl<'a> Write for &'a TcpStreamimpl<'a> Read for &'a TcpStreamimpl Read for TcpStream

fn as_tcp(&self) -> &TcpStream

Get a borrow of the TcpStream