Trait websocket::stream::sync::Splittable

pub trait Splittable {
    type Reader: Read;
    type Writer: Write;
    fn split(self) -> Result<(Self::Reader, Self::Writer)>;
}

some streams can be split up into separate reading and writing components TcpStream is an example. This trait marks this ability so one can split up the client into two parts.

Notice however that this is not possible to do with SSL.

Associated Types

type Reader: Read

The reading component of this type

type Writer: Write

The writing component of this type

Required Methods

fn split(self) -> Result<(Self::Reader, Self::Writer)>

Split apart this type into a reading and writing component.