Struct websocket::WebSocketClient

pub struct WebSocketClient {
    // some fields omitted
}

Represents a WebSocket client. To use WebSocketClient, you must create one using either WebSocketClient::connect(), which is used for writing clients, or WebSocketClient::from_stream(), which creates a WebSocketClient from a TcpStream (typically used in a server).

An example client application:

use websocket::WebSocketClient;
use websocket::handshake::WebSocketRequest;
 
let request = WebSocketRequest::new("ws://127.0.0.1:1234", "myProtocol").unwrap();
let key = request.key().unwrap();
let mut client = WebSocketClient::connect(&request).unwrap();
let response = client.receive_handshake_response().unwrap();
 
if !response.is_successful(key) {
    // Handshake failed!
}
 
// Now we can send and receive messages
let receiver = client.receiver();
let mut sender = client.sender();
 
// ...

Methods

impl WebSocketClient

fn connect(request: &WebSocketRequest) -> IoResult<WebSocketClient>

Connect to the WebSocket server using the given request. Use WebSocketRequest::new() to create a request for use with this function.

fn from_stream(stream: TcpStream, mask: bool) -> WebSocketClient

Creates a new WebSocketClient from a given TcpStream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false.

fn stream(&self) -> TcpStream

Returns a copy of the underlying TcpStream for this WebSocketClient.

fn peer_name(&mut self) -> IoResult<SocketAddr>

Returns the socket address of the remote peer of this TCP connection.

fn socket_name(&mut self) -> IoResult<SocketAddr>

Returns the socket address of the local half of this TCP connection.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn receive_handshake_response(&mut self) -> IoResult<WebSocketResponse>

Reads a response that was sent to this client. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_response(&mut self, response: WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn close_send(&mut self) -> IoResult<()>

Closes the sender for this WebSocketClient. This method will close the message sending portion of this client, causing all pending and future sends to immediately return with an error. This affects all WebSocketSenders for the client, and any copies of the underlying stream will be unable to write.

Note that you should send a WebSocketMessage:Close message to the remote endpoint before calling this method.

fn receiver(&self) -> WebSocketReceiver

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.

fn close_receive(&mut self) -> IoResult<()>

Closes the receiver for this WebSocketClient. This method will close the message receiving portion of this client, causing all pending and future receives to immediately return with an error. This affects all WebSocketReceivers for the client, and any copies of the underlying stream will be unable to read.

Note that you should send a WebSocketMessage:Close message to the remote endpoint before calling this method.

impl WebSocketClient

fn connect(request: &WebSocketRequest) -> IoResult<WebSocketClient>

Connect to the WebSocket server using the given request. Use WebSocketRequest::new() to create a request for use with this function.

fn from_stream(stream: TcpStream, mask: bool) -> WebSocketClient

Creates a new WebSocketClient from a given TcpStream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false.

fn stream(&self) -> TcpStream

Returns a copy of the underlying TcpStream for this WebSocketClient.

fn peer_name(&mut self) -> IoResult<SocketAddr>

Returns the socket address of the remote peer of this TCP connection.

fn socket_name(&mut self) -> IoResult<SocketAddr>

Returns the socket address of the local half of this TCP connection.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn receive_handshake_response(&mut self) -> IoResult<WebSocketResponse>

Reads a response that was sent to this client. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_response(&mut self, response: WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn close_send(&mut self) -> IoResult<()>

Closes the sender for this WebSocketClient. This method will close the message sending portion of this client, causing all pending and future sends to immediately return with an error. This affects all WebSocketSenders for the client, and any copies of the underlying stream will be unable to write.

Note that you should send a WebSocketMessage:Close message to the remote endpoint before calling this method.

fn receiver(&self) -> WebSocketReceiver

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.

fn close_receive(&mut self) -> IoResult<()>

Closes the receiver for this WebSocketClient. This method will close the message receiving portion of this client, causing all pending and future receives to immediately return with an error. This affects all WebSocketReceivers for the client, and any copies of the underlying stream will be unable to read.

Note that you should send a WebSocketMessage:Close message to the remote endpoint before calling this method.