Struct websocket::WebSocketClient

pub struct WebSocketClient<S: Stream + Clone> {
    // some fields omitted
}

Represents a WebSocket client. To use WebSocketClient, you must create one using WebSocketClient::new(). For a client, you can use TcpStream::connect() to connect to the server, then call WebSocketClient::new() passing the resultant stream.

use std::io::TcpStream;
use websocket::WebSocketClient;
use websocket::handshake::WebSocketRequest;
 
let request = WebSocketRequest::new("ws://127.0.0.1:1234", ["myProtocol"].as_slice()).unwrap();
let key = request.key().unwrap();

// We can get the hostname from the request
let stream = TcpStream::connect(request.host().unwrap().as_slice()).unwrap();

let mut client = WebSocketClient::new(stream, true);
let _ = client.send_handshake_request(&request);
let response = client.receive_handshake_response().unwrap();
 
if !response.is_successful(key) {
    // Handshake failed!
}
 
// Now we can send and receive messages
let receiver = client.receiver();
let mut sender = client.sender();
 
// ...

Methods

impl<S: Stream + Clone> WebSocketClient<S>

fn new(stream: S, mask: bool) -> WebSocketClient<S>

Creates a new WebSocketClient from a given cloneable stream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false.

Nothing is sent to or read from the stream during the conversion.

fn stream(&self) -> S

Returns a copy of the underlying S for this WebSocketClient. Note that writing to this stream will likely invalidate the WebSocket data stream.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn receive_handshake_response(&mut self) -> IoResult<WebSocketResponse>

Reads a response that was sent to this client. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_request(&mut self, request: &WebSocketRequest) -> IoResult<()>

Sends the specified WebSocketRequest to the remote endpoint. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_response(&mut self, response: &WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender<S>

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn receiver(&self) -> WebSocketReceiver<S>

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.

impl<S: Stream + Clone> WebSocketClient<S>

fn new(stream: S, mask: bool) -> WebSocketClient<S>

Creates a new WebSocketClient from a given cloneable stream. The mask parameter determines whether or not messages send to the remote endpoint will be masked. If the client is connecting to a remote endpoint, set mask to true. If the client is the remote endpoint (and therefore, the server is the local endpoint), set mask to false.

Nothing is sent to or read from the stream during the conversion.

fn stream(&self) -> S

Returns a copy of the underlying S for this WebSocketClient. Note that writing to this stream will likely invalidate the WebSocket data stream.

fn receive_handshake_request(&mut self) -> IoResult<WebSocketRequest>

Reads a request from this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn receive_handshake_response(&mut self) -> IoResult<WebSocketResponse>

Reads a response that was sent to this client. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_request(&mut self, request: &WebSocketRequest) -> IoResult<()>

Sends the specified WebSocketRequest to the remote endpoint. Only to be used if the server is the remote endpoint and the client is the local endpoint.

fn send_handshake_response(&mut self, response: &WebSocketResponse) -> IoResult<()>

Sends the specified WebSocketResponse to this client. Only to be used if the server is the local endpoint and the client is the remote endpoint.

fn sender(&self) -> WebSocketSender<S>

Returns a WebSocketSender from this client. Used to transmit data to the remote endpoint, that is, to the server if WebSocketClient::connect() has been used, or to this client otherwise.

fn receiver(&self) -> WebSocketReceiver<S>

Returns a WebSocketReceiver from this client. Used to receive data from the remote endpoint, that is, from the server if WebSocketClient::connect() has been used, or from this client otherwise.