Trait websocket::ws::dataframe::DataFrame

pub trait DataFrame {
    fn is_last(&self) -> bool;

    fn opcode(&self) -> u8;

    fn reserved(&self) -> &[bool; 3];

    fn size(&self) -> usize;

    fn write_payload(&self, socket: &mut Write) -> WebSocketResult<()>;

    fn take_payload(self) -> Vec<u8>;

    fn frame_size(&self, masked: bool) -> usize { ... }

    fn write_to(&self, writer: &mut Write, mask: bool) -> WebSocketResult<()> { ... }
}

A generic DataFrame. Every dataframe should be able to provide these methods. (If the payload is not known in advance then rewrite the write_payload method)

Required Methods

fn is_last(&self) -> bool

Is this dataframe the final dataframe of the message?

fn opcode(&self) -> u8

What type of data does this dataframe contain?

fn reserved(&self) -> &[bool; 3]

Reserved bits of this dataframe

fn size(&self) -> usize

How long (in bytes) is this dataframe's payload

fn write_payload(&self, socket: &mut Write) -> WebSocketResult<()>

Write the payload to a writer

Important traits for Vec<u8>
impl Write for Vec<u8>

fn take_payload(self) -> Vec<u8>

Takes the payload out into a vec

Provided Methods

fn frame_size(&self, masked: bool) -> usize

Get's the size of the entire dataframe in bytes, i.e. header and payload.

fn write_to(&self, writer: &mut Write, mask: bool) -> WebSocketResult<()>

Writes a DataFrame to a Writer.