Trait websocket::ws::message::Message

pub trait Message<'a, F>: Sized where
    F: DataFrame, {
    type DataFrameIterator: Iterator<Item = F>;
    fn from_dataframes<D>(frames: Vec<D>) -> WebSocketResult<Self>
    where
        D: DataFrame;

    fn dataframes(&'a self) -> Self::DataFrameIterator;
}

A trait for WebSocket messages

Associated Types

type DataFrameIterator: Iterator<Item = F>

The iterator type returned by dataframes

Required Methods

fn from_dataframes<D>(frames: Vec<D>) -> WebSocketResult<Self> where
D: DataFrame,

Attempt to form a message from a slice of data frames.

fn dataframes(&'a self) -> Self::DataFrameIterator

Turns this message into an iterator over data frames