Trait websocket::ws::message::Message

pub trait Message {
    fn serialize(
        &self, 
        &mut dyn Write, 
        masked: bool
    ) -> Result<(), WebSocketError>;

    fn message_size(&self, masked: bool) -> usize;

    fn from_dataframes<D>(
        frames: Vec<D, Global>
    ) -> Result<Self, WebSocketError>
    where
        D: DataFrame;
}

A trait for WebSocket messages

Required methods

fn serialize(&self, &mut dyn Write, masked: bool) -> Result<(), WebSocketError>

Writes this message to the writer

fn message_size(&self, masked: bool) -> usize

Returns how many bytes this message will take up

fn from_dataframes<D>(frames: Vec<D, Global>) -> Result<Self, WebSocketError> where
D: DataFrame,

Attempt to form a message from a series of data frames