Trait websocket::ws::receiver::Receiver

pub trait Receiver: Sized {
    type F: DataFrame;
    type M: Message;
    fn recv_dataframe<R>(&mut self, reader: &mut R) -> WebSocketResult<Self::F>
    where
        R: Read;
    fn recv_message_dataframes<R>(
        &mut self, 
        reader: &mut R
    ) -> WebSocketResult<Vec<Self::F>>
    where
        R: Read;

    fn incoming_dataframes<'a, R>(
        &'a mut self, 
        reader: &'a mut R
    ) -> DataFrameIterator<'a, Self, R>
    where
        R: Read,
    { ... }
    fn recv_message<'m, R>(
        &mut self, 
        reader: &mut R
    ) -> WebSocketResult<Self::M>
    where
        R: Read,
    { ... }
    fn incoming_messages<'a, R>(
        &'a mut self, 
        reader: &'a mut R
    ) -> MessageIterator<'a, Self, R>
    where
        R: Read,
    { ... }
}

A trait for receiving data frames and messages.

Associated Types

type F: DataFrame

The type of dataframe that incoming messages will be serialized to.

type M: Message

The type of message that incoming messages will be serialized to.

Required Methods

fn recv_dataframe<R>(&mut self, reader: &mut R) -> WebSocketResult<Self::F> where
R: Read,

Reads a single data frame from this receiver.

fn recv_message_dataframes<R>(
&mut self,
reader: &mut R
) -> WebSocketResult<Vec<Self::F>> where
R: Read,

Returns the data frames that constitute one message.

Provided Methods

fn incoming_dataframes<'a, R>(
&'a mut self,
reader: &'a mut R
) -> DataFrameIterator<'a, Self, R> where
R: Read,

Returns an iterator over incoming data frames.

fn recv_message<'m, R>(&mut self, reader: &mut R) -> WebSocketResult<Self::M> where
R: Read,

Reads a single message from this receiver.

fn incoming_messages<'a, R>(
&'a mut self,
reader: &'a mut R
) -> MessageIterator<'a, Self, R> where
R: Read,

Returns an iterator over incoming messages.